+++ backend/mora/service/engagement.py
@@ -47,6 +47,9 @@ class EngagementRequestHandler(handlers.OrgFunkRequestHandler):
         validator.is_date_range_in_org_unit_range(org_unit, valid_from,
                                                   valid_to)
 
+        func_id = util.get_uuid(req, required=False) or str(uuid.uuid4())
+        bvn = util.checked_get(req, mapping.USER_KEY, func_id)
+
         primary = req.get(mapping.PRIMARY)
 
         if primary:
@@ -63,8 +66,6 @@ class EngagementRequestHandler(handlers.OrgFunkRequestHandler):
                                                      mapping.ENGAGEMENT_TYPE,
                                                      required=True)
 
-        bvn = util.checked_get(req, mapping.USER_KEY, str(uuid.uuid4()))
-
         payload = common.create_organisationsfunktion_payload(
             funktionsnavn=mapping.ENGAGEMENT_KEY,
             primær=primary,
@@ -75,11 +76,12 @@ class EngagementRequestHandler(handlers.OrgFunkRequestHandler):
             tilknyttedeorganisationer=[org_uuid],
             tilknyttedeenheder=[org_unit_uuid],
             funktionstype=engagement_type_uuid,
-            opgaver=[{'uuid': job_function_uuid}] if job_function_uuid else []
+            opgaver=[{'uuid': job_function_uuid}] if job_function_uuid else [],
+            integration_data=req.get(mapping.INTEGRATION_DATA),
         )
 
         self.payload = payload
-        self.uuid = util.get_uuid(req, required=False)
+        self.uuid = func_id
 
     def prepare_edit(self, req: dict):
         engagement_uuid = util.get_uuid(req)
@@ -127,6 +129,12 @@ class EngagementRequestHandler(handlers.OrgFunkRequestHandler):
             mapping.ORG_FUNK_GYLDIGHED_FIELD,
             {'gyldighed': "Aktiv"}
         ))
+
+        if mapping.USER_KEY in data:
+            update_fields.append((
+                mapping.ORG_FUNK_EGENSKABER_FIELD,
+                {'brugervendtnoegle': data[mapping.USER_KEY]},
+            ))
 
         if mapping.JOB_FUNCTION in data:
             update_fields.append((