+++ backend/tests/test_integration_engagement.py
@@ -49,7 +49,7 @@ class Tests(util.LoRATestCase):
                                            json=payload)
 
         expected = {
-            "livscykluskode": "Opstaaet",
+            "livscykluskode": "Importeret",
             "tilstande": {
                 "organisationfunktiongyldighed": [
                     {
@@ -172,7 +172,7 @@ class Tests(util.LoRATestCase):
                                                json=payload)
 
         expected = {
-            "livscykluskode": "Opstaaet",
+            "livscykluskode": "Importeret",
             "tilstande": {
                 "organisationfunktiongyldighed": [
                     {
@@ -295,7 +295,7 @@ class Tests(util.LoRATestCase):
                                                json=payload)
 
         expected = {
-            "livscykluskode": "Opstaaet",
+            "livscykluskode": "Importeret",
             "tilstande": {
                 "organisationfunktiongyldighed": [
                     {
@@ -416,7 +416,7 @@ class Tests(util.LoRATestCase):
                                                json=payload)
 
         expected = {
-            "livscykluskode": "Opstaaet",
+            "livscykluskode": "Importeret",
             "tilstande": {
                 "organisationfunktiongyldighed": [
                     {
@@ -619,6 +619,7 @@ class Tests(util.LoRATestCase):
             "uuid": engagement_uuid,
             "data": {
                 "primary": True,
+                "user_key": "regnormsberiger",
                 "job_function": {
                     'uuid': "cac9c6a8-b432-4e50-b33e-e96f742d4d56"},
                 "engagement_type": {
@@ -744,9 +745,19 @@ class Tests(util.LoRATestCase):
                             "from_included": True,
                             "to_included": False,
                             "from": "2017-01-01 00:00:00+01",
-                            "to": "infinity"
+                            "to": "2018-04-01 00:00:00+02"
                         },
                         "brugervendtnoegle": "bvn",
+                        "funktionsnavn": "Engagement"
+                    },
+                    {
+                        "virkning": {
+                            "from_included": True,
+                            "to_included": False,
+                            "from": "2018-04-01 00:00:00+02",
+                            "to": "infinity"
+                        },
+                        "brugervendtnoegle": "regnormsberiger",
                         "funktionsnavn": "Engagement"
                     }
                 ],
@@ -1238,6 +1249,7 @@ class Tests(util.LoRATestCase):
                 "name": "Anders And",
                 "uuid": "53181ed2-f1de-4c4a-a8fd-ab358c2c454a",
             },
+            "user_key": "bvn",
             "uuid": "d000591f-8705-4324-897a-075e3623f37b",
         }
 
@@ -1771,6 +1783,7 @@ class Tests(util.LoRATestCase):
                             'name': 'Fedtmule',
                             'uuid': '6ee24785-ee9a-4502-81c2-7697009c9053',
                         },
+                        'user_key': '00000000-0000-0000-0000-000000000001',
                         'primary': True,
                         'uuid': engagementid,
                         'validity': {'from': '2017-12-01', 'to': '2017-12-31'},
@@ -1851,7 +1864,7 @@ class Tests(util.LoRATestCase):
             )
 
         expected = {
-            "livscykluskode": "Opstaaet",
+            "livscykluskode": "Importeret",
             "tilstande": {
                 "organisationfunktiongyldighed": [
                     {