+++ frontend/e2e-tests/MoOrganisationUnitITSystem.js
@@ -2,7 +2,7 @@ import VueSelector from 'testcafe-vue-selectors'
 import { baseURL, reset } from './support'
 
 fixture('MoOrganisationUnitITSystem')
-  .afterEach(reset)
+  .beforeEach(reset)
   .page(`${baseURL}/organisation/97337de5-6096-41f9-921e-5bed7a140d85`)
 
 // skip: test data lacks units with IT systems